+++ packages/charting/src/components/HorizontalBarChart/examples/HorizontalBarChart.Basic.Example.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import { HorizontalBarChart, IHorizontalBarChartProps } from '@uifabric/charting';
+import { HorizontalBarChart, IHorizontalBarChartProps } from '@uifabric/charting/lib/HorizontalBarChart';
 import { customizable } from 'office-ui-fabric-react/lib/Utilities';
-@customizable('StackedBarChartBasicExample', ['theme', 'styles'])
+@customizable('HorizontalBarChartBasicExample', ['theme', 'styles'])
 export class HorizontalBarChartBasicExample extends React.Component<IHorizontalBarChartProps, {}> {
   constructor(props: IHorizontalBarChartProps) {
     super(props);
