+++ packages/charting/src/components/StackedBarChart/StackedBarChartPage.tsx
@@ -3,8 +3,11 @@ import * as React from 'react';
 import { ComponentPage, ExampleCard, IComponentDemoPageProps, PropertiesTableSet } from '@uifabric/example-app-base';
 
 import { StackedBarChartBasicExample } from './examples/StackedBarChart.Basic.Example';
+import { StackedBarChartMultiExample } from './examples/StackedBarChart.Multi.Example';
 
 const StackedBarChartBasicExampleCode = require('!raw-loader!@uifabric/charting/src/components/StackedBarChart/examples/StackedBarChart.Basic.Example.tsx') as string;
+
+const StackedBarChartMultiExampleCode = require('!raw-loader!@uifabric/charting/src/components/StackedBarChart/examples/StackedBarChart.Multi.Example.tsx') as string;
 
 export class StackedBarChartPage extends React.Component<IComponentDemoPageProps, {}> {
   public render(): JSX.Element {
@@ -16,6 +19,9 @@ export class StackedBarChartPage extends React.Component<IComponentDemoPageProps
           <div>
             <ExampleCard title="StackedBarChart" code={StackedBarChartBasicExampleCode}>
               <StackedBarChartBasicExample />
+            </ExampleCard>
+            <ExampleCard title="StackedBarChart" code={StackedBarChartMultiExampleCode}>
+              <StackedBarChartMultiExample />
             </ExampleCard>
           </div>
         }
