+++ packages/charting/src/components/VerticalBarChart/examples/VerticalBarChart.Basic.Example.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
-import { VerticalBarChart, IVerticalBarChartProps } from '@uifabric/charting';
+import { VerticalBarChart, IVerticalBarChartProps } from '@uifabric/charting/lib/VerticalBarChart';
 import { customizable } from 'office-ui-fabric-react/lib/Utilities';
-@customizable('StackedBarChartBasicExample', ['theme', 'styles'])
+@customizable('VerticalBarChartBasicExample', ['theme', 'styles'])
 export class VerticalBarChartBasicExample extends React.Component<IVerticalBarChartProps, {}> {
   constructor(props: IVerticalBarChartProps) {
     super(props);
